Scripts/requirements.txt
Faker
mysql-connector-python
mariadb 
psycopg2-binary